+++ src/ErrorHandlerListenerDelegatorFactory.php
@@ -5,7 +5,7 @@ declare(strict_types=1);
 namespace Los\MezzioSwooleNewrelic;
 
 use Psr\Container\ContainerInterface;
-use Zend\Stratigility\Middleware\ErrorHandler;
+use Laminas\Stratigility\Middleware\ErrorHandler;
 
 class ErrorHandlerListenerDelegatorFactory
 {
@@ -13,9 +13,11 @@ class ErrorHandlerListenerDelegatorFactory
     {
         /* @var ErrorHandler $errorHandler */
         $errorHandler = $callback();
-        $errorHandler->attachListener(function($error) {
-            newrelic_notice_error($error);
-        });
+        if (extension_loaded('newrelic')) {
+            $errorHandler->attachListener(function ($error) {
+                newrelic_notice_error($error);
+            });
+        }
 
         return $errorHandler;
     }