+++ src/NewRelicMiddleware.php
@@ -8,13 +8,15 @@ use Psr\Http\Message\ResponseInterface as Response;
 use Psr\Http\Message\ServerRequestInterface as Request;
 use Psr\Http\Server\MiddlewareInterface;
 use Psr\Http\Server\RequestHandlerInterface;
-use Zend\Expressive\Router\RouteResult;
+use Mezzio\Router\RouteResult;
 
 class NewRelicMiddleware implements MiddlewareInterface
 {
     public function process(Request $request, RequestHandlerInterface $handler): Response
     {
-        $this->detectTransactionName($request);
+        if (extension_loaded('newrelic')) {
+            $this->detectTransactionName($request);
+        }
 
         return $handler->handle($request);
     }